Add a new property dedicated to the continuous deployment of the project.
Add a new Maven plugin to get the current commit hash from git and add
a configuration section to the source, jar and war plugins in order to
set additional properties into the MANIFEST.MF file: the git commit hash,
the build version and the build timestamp.
The git commit hash will be used by our CI to determine from which commit
hash the release of a stable version has to be done.