Bug 10296 The cause of the bug can be: * the LookAuroraHelper object isn't in the opened session, * the localized bundle in the LookAuroraHelper is null.
For the first point, it is not possible because the LookAuroraHelper is
initialized by the MainFrame.jsp that embodies the layout and the
TopBar.jsp is itself loaded from this layout by the Silverpeas layouting
mechanism.
Nevertheless, in order to circumvent such a problem, in case the
LookAuroraHelper isn't yet initialized in the TopBar.jsp, the LookAuroraHelper
object is now instantiated if not yet by a new method:
LookAuroraHelper#getLookHelper(HttpSession) and this method is invoked in the
different JSP of the Aurora Look instead of the
LookHelper#getLookHelper(HttpSession) method.

For the second point, the localization bundle is initialized as an attribute of
the LookAuroraHelper at its construction. So, null localization bundle isn't
possible. In any way, to fix this problem, the localization bundle is'nt more
set as an attribute of the LookAuroraHelper but it is computed each time it is
asked (and as there is a cache management in ResourceLocator of the different
bundles, it is similar to an attribute set).

My thought about the problem is that it can occurs when a session is
transparently serialized/deserialized by Wildfly. The LookHelper classes aren't
marked as Serializable and they currently cannot be marked as such because they
have non serializable attributes (for example, resource bundles aren't
serializable, Silverpeas services aren't serializable, ...). The correct way
to fix this is to refactor the LookHelper in order either to be serializable or
to be CDI session scoped beans.