Responding to @mmoqui requested changes on the previous PR
Responding to @mmoqui requested changes on the previous PR
In the XML resources, missing de translation of some values. I commented out in data.xml, but the missing is also in the others XMLs.

Just one point was ignored because you ask to change the key (not a label)  define by default (carousel) . There is a notion of lang but I don't understand why, because it is just to defined the key  used by default

This part :
> @@ -167,6 +239,14 @@
             <name>values</name>
             <value lang="fr">Liste##Carrousel</value>
         </parameter>
+        <parameter>
+            <name>values</name>
+            <value lang="en">List##Carrousel</value>
+        </parameter>
+        <parameter>
+            <name>values</name>
+            <value lang="de">List##Carrousel</value>
+        </parameter>
         <parameter>
             <name>default</name>
             <value lang="fr">carousel</value>